# Speed up estimates loading