--- a/zipline/pipeline/loaders/blaze/estimates.py
+++ b/zipline/pipeline/loaders/blaze/estimates.py
@@ -166,7 +166,6 @@ class BlazeSplitAdjustedEstimatesLoader(BlazeEstimatesLoader):
             for column_name in self._split_adjusted_column_names
             if column_name in requested_column_names
         ]
-
         raw = load_raw_data(
             assets,
             dates,