--- a/zipline/pipeline/loaders/earnings_estimates.py
+++ b/zipline/pipeline/loaders/earnings_estimates.py
@@ -25,8 +25,8 @@ from zipline.pipeline.loaders.base import PipelineLoader
 from zipline.utils.numpy_utils import datetime64ns_dtype, float64_dtype
 from zipline.pipeline.loaders.utils import (
     ffill_across_cols,
-    last_in_date_group
-)
+    last_in_date_group,
+    flat_last_in_date_group)
 
 
 INVALID_NUM_QTRS_MESSAGE = "Passed invalid number of quarters %s; " \
@@ -692,13 +692,8 @@ class EarningsEstimatesLoader(PipelineLoader):
         # Get a DataFrame indexed by date with a MultiIndex of columns of [
         # self.estimates.columns, normalized_quarters, sid], where each cell
         # contains the latest data for that day.
-        last_per_qtr = last_in_date_group(
-            self.estimates,
-            dates,
-            assets_with_data,
-            reindex=True,
-            extra_groupers=[NORMALIZED_QUARTERS],
-        )
+        last_per_qtr = flat_last_in_date_group(self.estimates, dates, [SID_FIELD_NAME, NORMALIZED_QUARTERS])
+        last_per_qtr = last_in_date_group(self.estimates, dates, assets_with_data, reindex=True, extra_groupers=[NORMALIZED_QUARTERS],)
         # Forward fill values for each quarter/sid/dataset column.
         ffill_across_cols(last_per_qtr, columns, self.name_map)
         # Stack quarter and sid into the index.