Speeds up estimates pipeline loader. **WIP** pending further improvements

**Some timings**

On this branch:
py.test tests/pipeline/test_quarters_estimates.py  42.86s user 1.74s system 98% cpu 45.408 total

On master:
py.test tests/pipeline/test_quarters_estimates.py  46.65s user 1.98s system 96% cpu 50.636 total